Delete or Update

+++ RestApiTodoApp/src/main/java/com/example/RestApiTodoApp/Response.java
@@ -1,22 +1,22 @@
 package com.example.RestApiTodoApp;
 
 public class Response{
-	private int id;
+	private Long id;
 	private boolean completed;
 	private String title;
 	private int userId;
 
-    public Response(int id,boolean completed,String title,int userId){
+    public Response(Long id,boolean completed,String title,int userId){
         this.id=id;
         this.completed=completed;
         this.title=title;
         this.userId=userId;
     }
-	public void setId(int id){
+	public void setId(Long id){
 		this.id = id;
 	}
 
-	public int getId(){
+	public Long getId(){
 		return id;
 	}
